Moved HTML files into includes folder and changed the paths in the relevant files.
